canvas scene graph: re-factored groupData is split into geode for strokes and photos

+++ src/libSGEntities/SceneState.cpp
@@ -81,7 +81,7 @@ void entity::SceneState::stripDataFrom(RootScene *scene)
         if (!cnv) continue;
         m_canvasDataFlags[i] = cnv->getVisibilityAll();
         m_canvasToolFlags[i] = cnv->getVisibilityFrameInternal();
-        for (int j=0; j < cnv->getNumPhotos(); ++j){
+        for (size_t j=0; j < cnv->getNumPhotos(); ++j){
             entity::Photo* photo = cnv->getPhotoFromIndex(j);
             if (!photo) continue;
             float t = photo->getTransparency();
