+++ src/tests/SceneStateTest.cpp
@@ -77,7 +77,7 @@ void SceneStateTest::testBasicApi()
     QString fname1 = "../../samples/ds-32.bmp";
     m_rootScene->addPhoto(fname1.toStdString());
     QCOMPARE((int)m_rootScene->getCanvasCurrent()->getNumChildren(), 1);
-    QCOMPARE(m_rootScene->getCanvasCurrent()->getNumPhotos(), 1);
+    QCOMPARE((int)m_rootScene->getCanvasCurrent()->getNumPhotos(), 1);
     state_stripped->stripDataFrom(m_rootScene);
     QVERIFY(!state_stripped->isEmpty());
     QCOMPARE(m_rootScene->getUserScene()->getPhotoFromIndex(m_rootScene->getCanvasCurrent(), 0)->getTransparency(), state_stripped->getPhotoTransparencies()[0]);
@@ -88,7 +88,7 @@ void SceneStateTest::testBasicApi()
     QString fname2 = "../../samples/test.bmp";
     m_rootScene->addPhoto(fname2.toStdString());
     QCOMPARE((int)m_rootScene->getCanvasCurrent()->getNumChildren(), 1);
-    QCOMPARE(m_rootScene->getCanvasCurrent()->getNumPhotos(), 1);
+    QCOMPARE((int)m_rootScene->getCanvasCurrent()->getNumPhotos(), 1);
     /* imitate mouse click within CanvasPhotoWidget delegate to trigger clickedTransparencyMinus signal */
     QSignalSpy spy1(m_canvasWidget->getCanvasDelegate(), SIGNAL(clickedTransparencyMinus(QModelIndex)));
     QTreeWidgetItem* itemCa = m_canvasWidget->topLevelItem(m_rootScene->getUserScene()->getCanvasIndex(m_rootScene->getCanvasCurrent()));
@@ -279,7 +279,7 @@ void SceneStateTest::testReadWrite()
     m_rootScene->setFilePath(filename.toStdString());
     QVERIFY(m_rootScene->isSetFilePath());
     QVERIFY(this->loadSceneFromFile());
-    QCOMPARE(m_rootScene->getUserScene()->getNumCanvases(), 3);
+    QCOMPARE((int)m_rootScene->getUserScene()->getNumCanvases(), 3);
     QVERIFY(m_rootScene->getUserScene()->getCanvas(index));
     QVERIFY(!m_rootScene->getUserScene()->getCanvas(index)->getVisibilityAll());
 
@@ -297,7 +297,7 @@ void SceneStateTest::testReadWrite()
 
 void SceneStateTest::testAddCanvas()
 {
-    QCOMPARE(m_rootScene->getUserScene()->getNumCanvases(), 3);
+    QCOMPARE((int)m_rootScene->getUserScene()->getNumCanvases(), 3);
     /* set 0-th canvas as invisible */
     int index = 0;
     this->onVisibilitySetCanvas(index);
@@ -314,7 +314,7 @@ void SceneStateTest::testAddCanvas()
 
     /* add new canvas to the scene */
     m_rootScene->addCanvas(osg::Vec3f(1,0,0), osg::Vec3f(1,1,1));
-    QCOMPARE(m_rootScene->getUserScene()->getNumCanvases(), 4);
+    QCOMPARE((int)m_rootScene->getUserScene()->getNumCanvases(), 4);
     QCOMPARE((int)state->getCanvasDataFlags().size(), 4);
     QCOMPARE((int)state->getCanvasToolFlags().size(), 4);
     QCOMPARE(state->getCanvasDataFlags()[0], false);
@@ -324,7 +324,7 @@ void SceneStateTest::testAddCanvas()
 
     /* do undo */
     m_undoStack->undo();
-    QCOMPARE(m_rootScene->getUserScene()->getNumCanvases(), 3);
+    QCOMPARE((int)m_rootScene->getUserScene()->getNumCanvases(), 3);
     QCOMPARE((int)state->getCanvasDataFlags().size(), 3);
     QCOMPARE((int)state->getCanvasToolFlags().size(), 3);
     QCOMPARE(state->getCanvasDataFlags()[0], false);
@@ -332,7 +332,7 @@ void SceneStateTest::testAddCanvas()
 
     /* do redo */
     m_undoStack->redo();
-    QCOMPARE(m_rootScene->getUserScene()->getNumCanvases(), 4);
+    QCOMPARE((int)m_rootScene->getUserScene()->getNumCanvases(), 4);
     QCOMPARE((int)state->getCanvasDataFlags().size(), 4);
     QCOMPARE((int)state->getCanvasToolFlags().size(), 4);
     QCOMPARE(state->getCanvasDataFlags()[0], false);
@@ -346,7 +346,7 @@ void SceneStateTest::testAddCanvas()
 
 void SceneStateTest::testAddPhoto()
 {
-    QCOMPARE(m_rootScene->getUserScene()->getNumCanvases(), 3);
+    QCOMPARE((int)m_rootScene->getUserScene()->getNumCanvases(), 3);
     /* set 0-th canvas as invisible */
     int index = 0;
     this->onVisibilitySetCanvas(index);
@@ -365,7 +365,7 @@ void SceneStateTest::testAddPhoto()
     QCOMPARE(m_rootScene->getCanvasCurrent(), m_canvas1.get());
     QString filename1 = "../../samples/ds-32.bmp";
     m_rootScene->addPhoto(filename1.toStdString());
-    QCOMPARE(m_canvas1->getNumPhotos(), 1);
+    QCOMPARE((int)m_canvas1->getNumPhotos(), 1);
 
     /* make sure the transparency is a part of scene states */
     QCOMPARE((int)state->getPhotoTransparencies().size(), 1);
@@ -376,7 +376,7 @@ void SceneStateTest::testAddPhoto()
     QCOMPARE(m_rootScene->getCanvasCurrent(), m_canvas2.get());
     QString filename2 = "../../samples/test.bmp";
     m_rootScene->addPhoto(filename2.toStdString());
-    QCOMPARE(m_canvas2->getNumPhotos(), 1);
+    QCOMPARE((int)m_canvas2->getNumPhotos(), 1);
 
     /* make sure the transparency is a part of scene states */
     QCOMPARE((int)state->getPhotoTransparencies().size(), 2);
@@ -412,7 +412,7 @@ void SceneStateTest::testDeleteCanvas()
     QCOMPARE(m_rootScene->getCanvasCurrent(), m_canvas0.get());
     QString filename0 = "../../samples/ds-32.bmp";
     m_rootScene->addPhoto(filename0.toStdString());
-    QCOMPARE(m_canvas0->getNumPhotos(), 1);
+    QCOMPARE((int)m_canvas0->getNumPhotos(), 1);
 
     /* take a bookmark */
     this->onBookmark();
@@ -428,18 +428,18 @@ void SceneStateTest::testDeleteCanvas()
     QCOMPARE(m_rootScene->getCanvasCurrent(), m_canvas0.get());
     QString filename1 = "../../samples/test.bmp";
     m_rootScene->addPhoto(filename1.toStdString());
-    QCOMPARE(m_canvas0->getNumPhotos(), 2);
+    QCOMPARE((int)m_canvas0->getNumPhotos(), 2);
     QCOMPARE((int)state->getPhotoTransparencies().size(), 2);
     /* check setting scene state does not fail */
     QVERIFY(m_rootScene->setSceneState(state));
 
     /* delete 0th canvas from scene */
     m_rootScene->editCanvasDelete(m_canvas0.get());
-    QCOMPARE(m_rootScene->getUserScene()->getNumCanvases(), 2);
+    QCOMPARE((int)m_rootScene->getUserScene()->getNumCanvases(), 2);
     QCOMPARE((int)state->getCanvasDataFlags().size(), 2);
     QCOMPARE((int) state->getCanvasToolFlags().size(), 2);
     QCOMPARE((int)state->getPhotoTransparencies().size(), 0);
-    QCOMPARE(m_rootScene->getUserScene()->getNumPhotos(), 0);
+    QCOMPARE((int)m_rootScene->getUserScene()->getNumPhotos(), 0);
     /* check setting scene state does not fail */
     QVERIFY(m_rootScene->setSceneState(state));
 
@@ -453,7 +453,7 @@ void SceneStateTest::testDeleteCanvas()
 
     /* perform redo */
     m_undoStack->redo();
-    QCOMPARE(m_rootScene->getUserScene()->getNumCanvases(), 2);
+    QCOMPARE((int)m_rootScene->getUserScene()->getNumCanvases(), 2);
     QCOMPARE((int)state->getCanvasDataFlags().size(), 2);
     QCOMPARE((int) state->getCanvasToolFlags().size(), 2);
     QCOMPARE((int)state->getPhotoTransparencies().size(), 0);
@@ -470,7 +470,7 @@ void SceneStateTest::testDeletePhoto()
     QCOMPARE(m_rootScene->getCanvasCurrent(), m_canvas0.get());
     QString filename0 = "../../samples/ds-32.bmp";
     m_rootScene->addPhoto(filename0.toStdString());
-    QCOMPARE(m_canvas0->getNumPhotos(), 1);
+    QCOMPARE((int)m_canvas0->getNumPhotos(), 1);
 
     /* take a bookmark */
     this->onBookmark();
@@ -486,7 +486,7 @@ void SceneStateTest::testDeletePhoto()
     QCOMPARE(m_rootScene->getCanvasCurrent(), m_canvas0.get());
     QString filename1 = "../../samples/test.bmp";
     m_rootScene->addPhoto(filename1.toStdString());
-    QCOMPARE(m_canvas0->getNumPhotos(), 2);
+    QCOMPARE((int)m_canvas0->getNumPhotos(), 2);
     QCOMPARE((int)state->getPhotoTransparencies().size(), 2);
     /* check setting scene state does not fail */
     QVERIFY(m_rootScene->setSceneState(state));
@@ -538,7 +538,7 @@ void SceneStateTest::testDeletePhoto()
 
     /* perform undo (return 1st photo with name test.bmp) */
     m_undoStack->undo();
-    QCOMPARE(m_rootScene->getUserScene()->getNumPhotos(), 1);
+    QCOMPARE((int)m_rootScene->getUserScene()->getNumPhotos(), 1);
     photo1 = m_rootScene->getUserScene()->getPhotoFromIndex(m_canvas0.get(), 0);
     QVERIFY(photo1);
     QCOMPARE(photo1->getName().c_str(), "Photo1");
